+++ src/main/java/elements/api/actions/ElementActions.java
@@ -4,7 +4,7 @@ import core.actions.Action;
 import core.actions.ActionCapability;
 import core.actions.ActionProfiles;
 import core.annotations.Internal;
-import core.engine.LocatorDescriptor;
+import elements.locator.LocatorDescriptor;
 import core.engine.UIEngine;
 import elements.api.UIElement;
 import elements.meta.ElementRole;